+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ## Added
 - Added cleanup mechanism of old compile reports (#2054)
 - Documentation on valid client types (#2015)
+- Added `compiler.json` option and `--json` compile flag to export structured compile data such as occurred errors (#1206)
+
 
 # v 2020.2 (2020-04-24) Changes in this release: